+++ libraries/AP_HAL_PX4/NSHShellStream.cpp
@@ -122,18 +122,18 @@ int16_t NSHShellStream::read()
     return -1;
 }
 
-int16_t NSHShellStream::available()
+uint32_t NSHShellStream::available()
 {
-    int ret = 0;
+    uint32_t ret = 0;
     if (ioctl(shell_stdin, FIONREAD, (unsigned long)&ret) == OK) {
         return ret;
     }
     return 0;
 }
 
-int16_t NSHShellStream::txspace()
+uint32_t NSHShellStream::txspace()
 {
-    int ret = 0;
+    uint32_t ret = 0;
     if (ioctl(shell_stdout, FIONWRITE, (unsigned long)&ret) == OK) {
         return ret;
     }
