+++ libraries/AP_HAL_VRBRAIN/UARTDriver.cpp
@@ -230,7 +230,7 @@ bool VRBRAINUARTDriver::tx_pending() { return false; }
 /*
   return number of bytes available to be read from the buffer
  */
-int16_t VRBRAINUARTDriver::available()
+uint32_t VRBRAINUARTDriver::available()
 { 
 	if (!_initialised) {
         try_initialise();
@@ -243,7 +243,7 @@ int16_t VRBRAINUARTDriver::available()
 /*
   return number of bytes that can be added to the write buffer
  */
-int16_t VRBRAINUARTDriver::txspace()
+uint32_t VRBRAINUARTDriver::txspace()
 { 
 	if (!_initialised) {
         try_initialise();
